+++ packages/stateful/lib/index.js
@@ -3,6 +3,12 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
+Object.defineProperty(exports, "StatefulHOC", {
+  enumerable: true,
+  get: function get() {
+    return _StatefulHOC.default;
+  }
+});
 Object.defineProperty(exports, "default", {
   enumerable: true,
   get: function get() {
@@ -14,6 +20,8 @@ exports.Status = void 0;
 var Status = _interopRequireWildcard(require("./Status"));
 
 exports.Status = Status;
+
+var _StatefulHOC = _interopRequireDefault(require("./StatefulHOC"));
 
 var _Stateful = _interopRequireDefault(require("./Stateful"));