+++ packages/stateful/src/index.js
@@ -1,3 +1,4 @@
 import * as Status from './Status';
 export { Status };
+export { default as StatefulHOC } from './StatefulHOC';
 export { default } from './Stateful';